% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{as_mapping}
\alias{as_mapping}
\title{Coerce objects to callable mappings}
\usage{
as_mapping(x, ...)
}
\arguments{
\item{x}{object to coerce}

\item{...}{reserved

Generic function. Methods for named vectors (chr -> vctr), dfs (df[[1]] -> df[[2]]), lists (l[[1]] -> l[[2]]), and functions (as is, but with no more than one required parameter).}
}
\description{
Coerce objects to callable mappings
}
