\name{distance.matrix}
\alias{distance.matrix}
\title{distance.matrix}
\usage{
  distance.matrix()
}
\value{
  A square matrix of distance multipliers
}
\description{
  Distance multipliers for distance conversion
}
\details{
  This is a conversion for common distance conversions such
  as kilometers to miles or feet to yards.
}
\examples{
\dontrun{
mapping:::distance.matrix()
}
}
\author{
  Jared P. Lander and Nicholas Galisinao
}

