% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyvalue.R
\name{keyvalue}
\alias{keyvalue}
\title{Key-value pairs}
\usage{
keyvalue(key, value)
}
\arguments{
\item{key}{vector of keys}

\item{value}{vector of values to which keys are mapped

Implemented to be compatible with `vctrs` and easily coercible to a mapper.}
}
\description{
Key-value pairs
}
