\name{map.options}
\alias{map.options}
\title{map.options}
\usage{
  map.options()
}
\value{
  A \code{\link{list}} of ggplot items.
}
\description{
  Common mapping options
}
\details{
  This builds a ggplot opts list that is commonly used for
  plotting.  Currently it only gives the default, perhaps
  I'll make it so users can pick and choose.
}
\author{
  Jared P. Lander
}

