\name{mapPalette}
\alias{mapPalette}

\title{
Color palette}
\description{
Returns different color palette}
\usage{
mapPalette(type, nclass = NULL )
}

\arguments{
  \item{type}{character value indicating the color palette}
  \item{nclass}{number of classes}
}

\value{A character vector with palettes.}
