% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping-package.R
\docType{package}
\name{mapping-package}
\alias{mapping}
\alias{mapping-package}
\title{mapping: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
