% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyvalue.R
\name{new_keyvalue}
\alias{new_keyvalue}
\title{Vectors of Key-Value Pairs}
\usage{
new_keyvalue(key = unspecified(), value = unspecified())
}
\arguments{
\item{key}{vector of keys}

\item{values}{vector of values

Length of the `key` and `value` vectors must be the same.
In order to maintain vctrs invariants, these are not valid mappings in this form.}
}
\description{
Vectors of Key-Value Pairs
}
