% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{new_mapping}
\alias{new_mapping}
\title{Low Level Constructor for Key-Value Mappings}
\usage{
new_mapping(keys, values)
}
\arguments{
\item{keys}{from the domain}

\item{values}{to which to map the keys

Right now mapping uses a named vector lookup, so only character vectors can be keys.
This will be generalized to vectors in the future.}
}
\description{
Low Level Constructor for Key-Value Mappings
}
\examples{
lookup <- new_mapping(c("m", "f", "u"), c("Male", "Female", NA))
x <- c("m", "f", "u", "f", "f", "m", "m")
lookup(x)
}
