\name{none}
\alias{none}
\title{none}
\usage{
  none(...)
}
\arguments{
  \item{\dots}{Anything.}
}
\value{
  \code{\link{NULL}} regardless of inputs.
}
\description{
  Just Null
}
\details{
  Returns NULL regardless of what is put in.
}
\examples{
none()
none(1)
none(mean)
}
\author{
  Jared P. Lander
}

