\name{popEU}
\alias{popEU}
\alias{popEUnuts2}
\docType{data}
\title{European population}
\description{European population for year 2018}

\usage{
data("popEU")
data("popEUnuts2")
}
\format{
  A data frame with 2252 observations on the following 5 variables.
  \describe{
    \item{\code{TIME}}{year}
    \item{\code{GEO}}{names}
    \item{\code{total}}{total}
    \item{\code{male}}{number of male}
    \item{\code{female}}{number of female}
  }
}

\source{
\url{https://ec.europa.eu/eurostat/data/database}}

\keyword{datasets}
