\name{popIT}
\alias{popIT}
\docType{data}
\title{Italian Population}
\description{Italian provincia population for year 2018}

\usage{data("popIT")}
\format{
  A data frame with 107 observations on the following 4 variables.
  \describe{
    \item{\code{ID}}{names}
    \item{\code{maschi}}{number of male}
    \item{\code{femmine}}{number of femal}
    \item{\code{totale}}{total}
  }
}

\source{Istat}

\keyword{datasets}
