\name{popUS}
\alias{popUS}
\docType{data}
\title{USA population}

\description{USA population for year 2019}

\usage{data("popUS")}
\format{
  A data frame with 52 observations on the following 2 variables.
  \describe{
    \item{\code{id}}{names}
    \item{\code{population}}{total population}
  }
}

\source{\url{https://www.census.gov/geographies/mapping-files/time-series/geo/carto-boundary-file.html}}

\keyword{datasets}
