\name{popWR}
\alias{popWR}
\docType{data}
\title{World population}
\description{Country pooulation for yeaar 2018}

\usage{data("popWR")}
\format{
  A data frame with 269 observations on the following 5 variables.
  \describe{
    \item{\code{country}}{a factor with countries}
    \item{\code{country_code}}{a factor with code}
    \item{\code{total}}{total}
    \item{\code{male}}{number of male}
    \item{\code{female}}{number of female}
  }
}


\source{World Bank}


\keyword{datasets}
