\name{within.distance.compute}
\alias{within.distance.compute}
\title{within.distance.compute}
\usage{
  within.distance.compute(dist.matrix, distance = 1)
}
\arguments{
  \item{dist.matrix}{Any matrix, though typically used with
  distance marices.}

  \item{distance}{The threshold to test against.}
}
\value{
  A vector with the counts of how many entries for each row
  are within the threshold.
}
\description{
  Count number of points within distance
}
\details{
  More Details coming later
}
\author{
  Jared P. Lander
}

