\name{mapping-package}
\alias{mappinng-package}
\docType{package}

\title{Worldwide, European, USA, and Italian static and interactive maps}

\description{
Maps are an inportant tool to visualise variables distribution acroos different spatial object. The mapping process require to link the data with coordinates and then generate the correspondent map. This package provide coordinates, linking and mapping functions for an automatic, flexible and easy approach of mapping workflows of differnt geographical statistical unit.Geographical coordinates are provided in the package and automatically linked with the input data to generate maps with internal provided functions or external functions.

}

\details{
  An introduction to \pkg{mappinng} package is provided in  vignette \href{../doc/a-journey-into-mapping.html}{A journey into mapping}.

  A graphical list of the available plots can be look at \href{../doc/type-of-plot-in-mapping.html}{Type of plots in mapping}

  Details on single country functions are also available are: \cr

  \href{../doc/mapping-world.html}{mapping World} \cr
  \href{../doc/mapping-european-union.html}{mapping European Union} \cr
  \href{../doc/mapping-italy.html}{mapping Italy} \cr
  \href{../doc/mapping-usa.html}{mapping USA} \cr



}



