\name{mapping-package}
\alias{mapping-package}
\alias{mapping}
\docType{package}
\title{
\packageTitle{mapping}
}
\description{
\packageDescription{mapping}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mapping}
\packageIndices{mapping}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{mapping}

Maintainer: \packageMaintainer{mapping}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
