% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{new_mapping}
\alias{new_mapping}
\title{Create a mapping that maps names to values}
\usage{
new_mapping(keyvals)
}
\description{
Create a mapping that maps names to values
}
\examples{
lookup <- new_mapper(c("m" = "Male", "f" = "Female", "u" = NA))
x <- c("m", "f", "u", "f", "f", "m", "m")
lookup(x)
}
